Do_Private_Type_Declaration
Do_Private_Type_Declaration registers the private type declaration
in the symbol table by obtaining its full view.
When the completion of the private type declaration is encountered
in the private part of the package the declaration is ignored because
it has already been registered in the symbol table when processing
the corresponding private type declaration.