The semantic analyser inserts declarations in sequences of statements.
Gnat2goto must accept them.
When the semantic analyser inserts object declarations with an
initialisation, it does not set the flag Has_Init_Expression.
The combined signed and unsigned integer to boolean
subprogram does not work correctly so a reverse loop which
increments the loop counter by -1 does not work.
Solved problem with reverse loop by writing a decrement
subprogram which decrements the loop counter by +1.