Suppress size too small messages when using gnat2goto.
These messages are generated and abort compilation even when
pragma Implicit_Packing should make the size acceptable.
This behaviour occurs when translating size representation clauses
in SPARK and gnat2goto.

The messages are supressed by add Gnatprove_Mode to the test which
causes the messages and subsequent compiler abort to be suppressed.