Initialisation of an array from a function call partially working.
Does not work for functions which return an unconstrained array type
or return an array which has non-static bounds.

This may be due to the goto return statement not assigning entire
arrays of these sorts.

Proposed solution is to declare an array object of the correct size,
determined from the body of the function, immediately before the call
of the function.  This will be an extra parameter of the function.
May be the type of the function should be changed from a dynamic array
to a *component_type and then it could return a pointer to its array
parameter?
In any case the result will be in the array object that is the extra
parameter.
In the case of unconstrained arrays further extra (out) parameters
will be introduced for the computed lower and upper bounds of each
dimension of the array.

To declare the extra objects that are required for functions
returning dynamic arrays, unfortunately, the current block will
have to be passed as an extra parameter to Do_Expression in
gnat2goto.