Solution proposed on previous commit will not work.
The array size returned by the function may not be known prior to the call of
the function and so cannot be passed as a parameter.  It could be passed as a
pointer and the array allocated via a call to malloc.

If malloc has to be used it would be better not to have extra parameters and
just return a pointer to the malloc'd array.  It should be possible to determine
its first, last and lenth from the type implicitly declared by the front-end to
represent the return type.

This commit has just the changes to Do_Function_Specification to cange the
result type of arrays to pointer to the array type.