Fix up forward processing of private and partial type declarations.
Private and incomplete type declarations are processed by gnat2goto
when the partial type declaration is encountered.  It uses the link
to the node of full type declaration which completes the partial
declaration.  This scheme of forward processing means that other type
declarations may appear between the partial type declaration and
its completion and may be used by the completion.
This change ensures that any itermediary declarations that occur
between a partial type declaration and its completion and are used
in its completion are processed by gnat2goto so that the type Irep
of the partial type is fully determined when it is processed.
It does this by recursing forward through the Atree processing any
intermediary declaration used by the completion of the private or
incomplete type.